// taurus-protect-sdk-typescript/src/internal/openapi/models/TgvalidatordMultiFactorSignaturesEntityType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Taurus-PROTECT APIs
 * Taurus-PROTECT APIs description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TgvalidatordMultiFactorSignaturesEntityType = {
    Request: 'REQUEST',
    WhitelistedAddress: 'WHITELISTED_ADDRESS',
    WhitelistedContract: 'WHITELISTED_CONTRACT'
} as const;
export type TgvalidatordMultiFactorSignaturesEntityType = typeof TgvalidatordMultiFactorSignaturesEntityType[keyof typeof TgvalidatordMultiFactorSignaturesEntityType];


export function instanceOfTgvalidatordMultiFactorSignaturesEntityType(value: any): boolean {
    for (const key in TgvalidatordMultiFactorSignaturesEntityType) {
        if (Object.prototype.hasOwnProperty.call(TgvalidatordMultiFactorSignaturesEntityType, key)) {
            if (TgvalidatordMultiFactorSignaturesEntityType[key as keyof typeof TgvalidatordMultiFactorSignaturesEntityType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TgvalidatordMultiFactorSignaturesEntityTypeFromJSON(json: any): TgvalidatordMultiFactorSignaturesEntityType {
    return TgvalidatordMultiFactorSignaturesEntityTypeFromJSONTyped(json, false);
}

export function TgvalidatordMultiFactorSignaturesEntityTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): TgvalidatordMultiFactorSignaturesEntityType {
    return json as TgvalidatordMultiFactorSignaturesEntityType;
}

export function TgvalidatordMultiFactorSignaturesEntityTypeToJSON(value?: TgvalidatordMultiFactorSignaturesEntityType | null): any {
    return value as any;
}

export function TgvalidatordMultiFactorSignaturesEntityTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): TgvalidatordMultiFactorSignaturesEntityType {
    return value as TgvalidatordMultiFactorSignaturesEntityType;
}